requirements.txt
python-whois
dnspython
requests
